Predict allele frequencies or mean phenotypes for a cline

Uses the cline parameters from your stanfit object (the mean of the posterior distribtution of each parameter) to predict the allele frequency (for genetic clines) or the mean phenotype value (for phenotypic clines) at the distance(s) along the transect specified by the user.

predict_cline(stanfit, distance)

Arguments

stanfit
A stanfit object holding your model results.

distance
The x value(s) (distance along the transect) at which to predict allele frequencies or mean phenotypes for the fitted cline. Must be a numeric vector.

Value

A data frame with two columns:

transectDist: the distances along the transect at which the cline equation was evaluated. Will be the same values as were supplied in the distance argument. Numeric.

p: the predict allele frequency or mean phenotype at each distance, according to the cline equation. Numeric.

Examples

# NOT RUN {
# Predict at one distance
predict_cline(yourStanfit, distance = 50)

# Predict at a range of distances
predict_cline(yourStanfit, distance = 0:100)

# }