Simulate genetic data from a cline

This function generates a dataframe with simulated genotypic data sampled from a genetic cline. The sampling sites, number of individuals, level of inbreeding, and cline parameters are supplied by the user. Cline parameters are flexible, and can model both sigmoid clines and stepped clines with introgresison tails.

sim_geno_cline(transect_distances, n_ind, Fis, decrease, center, width,
  pmin = 0, pmax = 1, deltaL = NULL, tauL = NULL, deltaR = NULL,
  tauR = NULL)

Arguments

transect_distances
The distances along the transect for the simulated sampling sites. A numeric vector.

n_ind
The number of diploid individuals sampled at each site. Either a single numeric value (for constant sampling), or a numeric vector equal in length to transect_distances.

Fis
The inbreeding coefficient, Fis, for each site. Must be between 0 and 1 (inclusive). Either a single numeric value (for constant inbreeding), or a numeric vector equal in length to transect_distances.

decrease
Is the cline decreasing in frequency? TRUE or FALSE.

center
The location of the cline center, in the same distance units as transect_distances. Numeric, must be greater than 0.

width
The width of the cline, in the same distance units as transect_distances. Numeric, must be greater than 0.

pmin, pmax
Optional. The minimum and maximum allele frequency values in the tails of the cline. Default values are 0 and 1, respectively. Must be between 0 and 1 (inclusive). Numeric.

deltaL, tauL
Optional delta and tau parameters which describe the left exponential tail. Must supply both to generate a tail. Default is NULL (no tails). Numeric. tauL must be between 0 and 1 (inclusive).

deltaR, tauR
Optional delta and tau parameters which describe the right exponential tail. Must supply both to generate a tail. Default is NULL (no tails). Numeric. tauR must be between 0 and 1 (inclusive).

Value

A data frame of simulated genetic data sampled from the cline. Columns are:

site: The site numbers, given sequentially starting at 1.

transectDist: The distance along the cline for each site.

cline.p: The expected allele frequency for each site, given its position on the cline.

cline.f: The expected coefficient of inbreeding for each site.

AA, Aa, aa: The simulated number of homozygotes and heterozygotes for each site.

N: The number of individuals sampled for each site.

emp.p: The observed allele frequency for each site (includes sampling error).

emp.f: The observed Fis for each site (includes sampling error).

Details

This function calls general_cline_eqn for each sampled point along the cline to generate the expected allele frequency at that point. The calculated allele frequency and provided Fis are then used to calculate the expected genotype frequencies for each site, according to the equations:

$$AA = p^2 + p(1-p)Fis$$ $$Aa = 2p(1-p)(1-Fis)$$ $$aa = (1-p)^2 + p(1-p)Fis$$

Sampled genotypes are then drawn from a multinomial distribution, with the expected genotype frequencies as the probabilities. From those sampled genotypes, empirical allele frequencies (emp.p) and empirical Fis estimates (emp.f) are calculated as:

$$emp.p = (2*AA + Aa)/N$$ $$emp.f = (Hexp - Hobs)/Hexp$$

where AA is the number of honozygotes of the focal allele, Aa is the number of heterozygotes, N is the number of sampled individuals, and Hexp and Hobs are the expected and observed heterozygosity. Fis values are corrected with correct_fis, such that empirical Fis values which are undefined or < 0 are corrected to 0.

Examples

# Simulate genotype data from a decreasing cline
# with center at 100, width of 30.
# Sites are 20 units apart, from 0 to 200.
# 20 individuals are sampled at each site.
# Inbreeding is constant at Fis = 0.1.

set.seed(123)
sim_geno_cline(transect_distance = seq(0,200,20),
               n_ind = 20, Fis = 0.1,
               decrease = TRUE,
               center = 100, width = 30)

#>    site transectDist cline.p cline.f AA Aa aa  N emp.p emp.f
#> 1     1            0   1.000     0.1 20  0  0 20 1.000 0.000
#> 2     2           20   1.000     0.1 20  0  0 20 1.000 0.000
#> 3     3           40   1.000     0.1 20  0  0 20 1.000 0.000
#> 4     4           60   0.995     0.1 19  1  0 20 0.975 0.000
#> 5     5           80   0.935     0.1 20  0  0 20 1.000 0.000
#> 6     6          100   0.500     0.1  6  6  8 20 0.450 0.394
#> 7     7          120   0.065     0.1  0  2 18 20 0.050 0.000
#> 8     8          140   0.005     0.1  0  0 20 20 0.000 0.000
#> 9     9          160   0.000     0.1  0  0 20 20 0.000 0.000
#> 10   10          180   0.000     0.1  0  0 20 20 0.000 0.000
#> 11   11          200   0.000     0.1  0  0 20 20 0.000 0.000

# Simulate genotype data from an increasing cline
# with center at 272, width of 91.
# The minimum and maximum allele frequencies
# are 0.07 and 0.98, respectively, and
# there is an introgression tail on the
# left side with deltaL = 29 and tauL = 0.8.

# Sites are 13 units apart, from 162 to 312.
# At each site, the number of individuals sampled
# is drawn from a random normal distribution with
# mean = 25 and sd = 5
# Inbreeding is constant at Fis = 0.

set.seed(123)
ind_sampling <- as.integer(rnorm(length(seq(162,312,12)), 25, 5))
sim_geno_cline(transect_distance = seq(162,312,12),
               n_ind = ind_sampling,
               Fis = 0, decrease = FALSE,
               center = 272, width = 91,
               pmin = 0.07, pmax = 0.98,
               deltaL = 29, tauL = 0.8)

#>    site transectDist cline.p cline.f AA Aa aa  N emp.p emp.f
#> 1     1          162   0.091       0  0  4 18 22 0.091 0.000
#> 2     2          174   0.100       0  0  2 21 23 0.043 0.000
#> 3     3          186   0.112       0  2  9 21 32 0.203 0.131
#> 4     4          198   0.128       0  1  7 17 25 0.180 0.051
#> 5     5          210   0.150       0  0  6 19 25 0.120 0.000
#> 6     6          222   0.182       0  2  7 24 33 0.167 0.236
#> 7     7          234   0.225       0  1  8 18 27 0.185 0.018
#> 8     8          246   0.290       0  0  8 10 18 0.222 0.000
#> 9     9          258   0.389       0  3 11  7 21 0.405 0.000
#> 10   10          270   0.505       0  4 14  4 22 0.500 0.000
#> 11   11          282   0.623       0 10 16  5 31 0.581 0.000
#> 12   12          294   0.729       0 15  8  3 26 0.731 0.218
#> 13   13          306   0.813       0 22  5  0 27 0.907 0.000