Simulate phenotypic data from a cline

This function generates a dataframe with simulated trait data sampled from a phenotypic cline. The sampling sites, number of individuals, phenotypic variance (sigma), and cline parameters are supplied by the user.

sim_pheno_cline(transect_distances, n_ind, sigma, decrease, center, width,
  pmin, pmax)

Arguments

transect_distances
The distances along the transect for the simulated sampling sites. A numeric vector.

n_ind
The number of individuals sampled at each site. Either a single numeric value (for constant sampling), or a numeric vector equal in length to transect_distances.

sigma
The simulated phenotypic variance, given in terms of standard deviation for a normal distribution. Must be numeric and > 0. Either a single numeric value (for constant variance), or a numeric vector equal in length to transect_distances.

decrease
Is the cline decreasing in frequency? TRUE or FALSE.

center
The location of the cline center, in the same distance units as transect_distances. Numeric, must be greater than 0.

width
The width of the cline, in the same distance units as transect_distances. Numeric, must be greater than 0.

pmin, pmax
The minimum and maximum mean phenotypic trait value in the tails of the cline. Numeric.

Value

A data frame of simulated phenotypic data sampled from the cline. Each row representes a single individual. Columns are:

site: The site numbers, given sequentially starting at 1.

transectDist: The distance along the cline for each individual

traitValue: The simulated trait value for each individual.

Details

This function calls general_cline_eqn for each sampled point along the cline to generate the expected mean phenotype at that point.

For each site the specified number of individuals are simulated. Trait values are drawn from a normal distribution, with mean equal to the expected mean phenotype and standard devitation equal to the provided sigma.

Examples

# Simulate phenotype data from a decreasing cline
# with center at 100, width of 30, pmin of 10 and pmax of 40.
# Sites are 20 units apart, from 0 to 200.
# 20 individuals are sampled at each site.
# Variance is constant at sigma = 6.

set.seed(123)
sim_pheno_cline(transect_distance = seq(0,200,20),
               n_ind = 20, sigma = 6,
               decrease = TRUE,
               center = 100, width = 30,
               pmin = 10, pmax = 40)

#>     site transectDist traitValue
#> 1      1            0 36.6370975
#> 2      1            0 38.6188865
#> 3      1            0 49.3522013
#> 4      1            0 40.4230018
#> 5      1            0 40.7756778
#> 6      1            0 50.2903413
#> 7      1            0 42.7654486
#> 8      1            0 32.4095840
#> 9      1            0 35.8788343
#> 10     1            0 37.3259796
#> 11     1            0 47.3444422
#> 12     1            0 42.1588344
#> 13     1            0 42.4045801
#> 14     1            0 40.6640477
#> 15     1            0 36.6649046
#> 16     1            0 50.7214302
#> 17     1            0 42.9870543
#> 18     1            0 28.2002485
#> 19     1            0 44.2080868
#> 20     1            0 37.1632030
#> 21     2           20 33.5923585
#> 22     2           20 38.6914513
#> 23     2           20 33.8432741
#> 24     2           20 35.6259534
#> 25     2           20 36.2490651
#> 26     2           20 29.8791409
#> 27     2           20 45.0260230
#> 28     2           20 40.9195395
#> 29     2           20 33.1704791
#> 30     2           20 47.5221903
#> 31     2           20 42.5580861
#> 32     2           20 38.2288718
#> 33     2           20 45.3700547
#> 34     2           20 45.2681017
#> 35     2           20 44.9287872
#> 36     2           20 44.1311423
#> 37     2           20 43.3228067
#> 38     2           20 39.6278305
#> 39     2           20 38.1635248
#> 40     2           20 37.7164747
#> 41     3           40 35.8216976
#> 42     3           40 38.7424358
#> 43     3           40 32.3975614
#> 44     3           40 53.0036753
#> 45     3           40 47.2377115
#> 46     3           40 33.2512880
#> 47     3           40 37.5726305
#> 48     3           40 37.1900074
#> 49     3           40 44.6697302
#> 50     3           40 39.4897251
#> 51     3           40 41.5098506
#> 52     3           40 39.8186590
#> 53     3           40 39.7327168
#> 54     3           40 48.2015532
#> 55     3           40 38.6353136
#> 56     3           40 49.0887631
#> 57     3           40 30.6974227
#> 58     3           40 43.4976220
#> 59     3           40 40.7330650
#> 60     3           40 41.2855889
#> 61     4           60 42.1336943
#> 62     4           60 36.8419167
#> 63     4           60 37.8566131
#> 64     4           60 33.7444051
#> 65     4           60 33.4251101
#> 66     4           60 41.6770293
#> 67     4           60 42.5451161
#> 68     4           60 40.1738828
#> 69     4           60 45.3894622
#> 70     4           60 52.1563655
#> 71     4           60 36.9096704
#> 72     4           60 26.0008442
#> 73     4           60 45.8902886
#> 74     4           60 35.6006528
#> 75     4           60 35.7278057
#> 76     4           60 46.0092856
#> 77     4           60 38.1472194
#> 78     4           60 32.5315511
#> 79     4           60 40.9436783
#> 80     4           60 39.0225092
#> 81     5           80 38.0855100
#> 82     5           80 40.3626073
#> 83     5           80 35.8269647
#> 84     5           80 41.9171842
#> 85     5           80 36.7280056
#> 86     5           80 40.0416167
#> 87     5           80 44.6319590
#> 88     5           80 40.6620139
#> 89     5           80 36.0953354
#> 90     5           80 44.9437706
#> 91     5           80 44.0119481
#> 92     5           80 41.3413067
#> 93     5           80 39.4833153
#> 94     5           80 34.2834885
#> 95     5           80 46.2148396
#> 96     5           80 34.4493674
#> 97     5           80 51.1749229
#> 98     5           80 47.2465887
#> 99     5           80 36.6367228
#> 100    5           80 31.8923995
#> 101    6          100 20.7375606
#> 102    6          100 26.5413023
#> 103    6          100 23.5198487
#> 104    6          100 22.9147444
#> 105    6          100 19.2902886
#> 106    6          100 24.7298337
#> 107    6          100 20.2905732
#> 108    6          100 14.9923484
#> 109    6          100 22.7186409
#> 110    6          100 30.5139797
#> 111    6          100 21.5479182
#> 112    6          100 28.6477859
#> 113    6          100 15.2927038
#> 114    6          100 24.6666282
#> 115    6          100 28.1164432
#> 116    6          100 26.8069202
#> 117    6          100 25.6340572
#> 118    6          100 21.1557640
#> 119    6          100 19.9017739
#> 120    6          100 18.8552273
#> 121    7          120 12.6549547
#> 122    7          120  6.2642274
#> 123    7          120  9.0057304
#> 124    7          120 10.4125219
#> 125    7          120 23.0122471
#> 126    7          120  8.0373757
#> 127    7          120 13.3613945
#> 128    7          120 12.4168402
#> 129    7          120  6.1779353
#> 130    7          120 11.5212266
#> 131    7          120 20.6163802
#> 132    7          120 14.6580994
#> 133    7          120 12.1964726
#> 134    7          120  9.4140941
#> 135    7          120 -0.3704083
#> 136    7          120 18.7370984
#> 137    7          120  3.1852346
#> 138    7          120 16.3887601
#> 139    7          120 23.4036965
#> 140    7          120  3.2857161
#> 141    8          140 14.3548486
#> 142    8          140  8.5709577
#> 143    8          140  0.7112776
#> 144    8          140  1.0561367
#> 145    8          140  0.5349255
#> 146    8          140  6.9587035
#> 147    8          140  1.3736091
#> 148    8          140 14.2716432
#> 149    8          140 22.7447962
#> 150    8          140  2.4219597
#> 151    8          140 14.8705757
#> 152    8          140 14.7583960
#> 153    8          140 12.1373581
#> 154    8          140  4.0938829
#> 155    8          140  9.4274269
#> 156    8          140  8.4617706
#> 157    8          140 13.5220798
#> 158    8          140  7.9095100
#> 159    8          140 16.0059829
#> 160    8          140  7.8966574
#> 161    9          160 16.3263293
#> 162    9          160  3.7149985
#> 163    9          160  2.4491290
#> 164    9          160 29.4563001
#> 165    9          160  7.5089150
#> 166    9          160 11.7994261
#> 167    9          160 13.8294785
#> 168    9          160  7.1073767
#> 169    9          160 13.1112328
#> 170    9          160 12.2238477
#> 171    9          160  8.7177775
#> 172    9          160 10.4018187
#> 173    9          160  9.8056570
#> 174    9          160 22.7807719
#> 175    9          160  5.5620439
#> 176    9          160  3.4340829
#> 177    9          160 10.2367909
#> 178    9          160 11.8729450
#> 179    9          160 12.6292014
#> 180    9          160  7.2598685
#> 181   10          180  3.6207425
#> 182   10          180 17.5798103
#> 183   10          180  7.9027969
#> 184   10          180  4.8076221
#> 185   10          180  8.5830218
#> 186   10          180  8.8176439
#> 187   10          180 16.6602210
#> 188   10          180 10.5091230
#> 189   10          180 14.5250220
#> 190   10          180  7.0049472
#> 191   10          180 11.2873711
#> 192   10          180  8.0525838
#> 193   10          180 10.5682004
#> 194   10          180  4.6285191
#> 195   10          180  2.1358901
#> 196   10          180 21.9839796
#> 197   10          180 13.6049522
#> 198   10          180  2.4930711
#> 199   10          180  6.3337038
#> 200   10          180  2.8878187
#> 201   11          200 23.1929107
#> 202   11          200 17.8745264
#> 203   11          200  8.4091782
#> 204   11          200 13.2592129
#> 205   11          200  7.5140089
#> 206   11          200  7.1425672
#> 207   11          200  5.2684316
#> 208   11          200  6.4323450
#> 209   11          200 19.9054934
#> 210   11          200  9.6758798
#> 211   11          200 10.7155200
#> 212   11          200 11.4621732
#> 213   11          200 17.3949039
#> 214   11          200  6.9036656
#> 215   11          200  4.0450057
#> 216   11          200 20.0542302
#> 217   11          200  7.3530693
#> 218   11          200  5.6616528
#> 219   11          200  2.5824099
#> 220   11          200  2.2917543

# Simulate phenotype data from an increasing cline
# with center at 272, width of 91.
# The minimum and maximum mean phenotype values
# are 32 and 49, respectively.

# Sites are 13 units apart, from 162 to 312.
# At each site, the number of individuals sampled
# is drawn from a random normal distribution with
# mean = 25 and sd = 5
# variance is constant at sigma = 4.67.

set.seed(123)
ind_sampling <- as.integer(rnorm(length(seq(162,312,12)), 25, 5))
sim_pheno_cline(transect_distance = seq(162,312,12),
               n_ind = ind_sampling,
               sigma = 4.67, decrease = FALSE,
               center = 272, width = 91,
               pmin = 32, pmax = 49)

#>     site transectDist traitValue
#> 1      1          162   32.65089
#> 2      1          162   29.53823
#> 3      1          162   40.47889
#> 4      1          162   34.45897
#> 5      1          162   22.94990
#> 6      1          162   35.40934
#> 7      1          162   29.92607
#> 8      1          162   27.14727
#> 9      1          162   31.11606
#> 10     1          162   27.34257
#> 11     1          162   28.73008
#> 12     1          162   29.21507
#> 13     1          162   24.25715
#> 14     1          162   36.04647
#> 15     1          162   32.85026
#> 16     1          162   26.81891
#> 17     1          162   37.98932
#> 18     1          162   34.12559
#> 19     1          162   30.75602
#> 20     1          162   36.31424
#> 21     1          162   36.23489
#> 22     1          162   35.97079
#> 23     2          174   35.44181
#> 24     2          174   34.81265
#> 25     2          174   31.93673
#> 26     2          174   30.79701
#> 27     2          174   30.44906
#> 28     2          174   28.98158
#> 29     2          174   31.25488
#> 30     2          174   26.31646
#> 31     2          174   42.35488
#> 32     2          174   37.86704
#> 33     2          174   26.98094
#> 34     2          174   30.34438
#> 35     2          174   30.04658
#> 36     2          174   35.86829
#> 37     2          174   31.83652
#> 38     2          174   33.40885
#> 39     2          174   32.09254
#> 40     2          174   32.02565
#> 41     2          174   38.61723
#> 42     2          174   31.17151
#> 43     2          174   39.30777
#> 44     2          174   24.99318
#> 45     2          174   34.95600
#> 46     3          186   32.95765
#> 47     3          186   33.38769
#> 48     3          186   34.15216
#> 49     3          186   30.03340
#> 50     3          186   30.82317
#> 51     3          186   27.62250
#> 52     3          186   27.37398
#> 53     3          186   33.79673
#> 54     3          186   34.47239
#> 55     3          186   32.62678
#> 56     3          186   36.68624
#> 57     3          186   41.95314
#> 58     3          186   30.08613
#> 59     3          186   21.59543
#> 60     3          186   37.07605
#> 61     3          186   29.06728
#> 62     3          186   29.16625
#> 63     3          186   37.16867
#> 64     3          186   31.04936
#> 65     3          186   26.67850
#> 66     3          186   33.22593
#> 67     3          186   31.73062
#> 68     3          186   32.40617
#> 69     3          186   34.17851
#> 70     3          186   30.64826
#> 71     3          186   35.38849
#> 72     3          186   31.34957
#> 73     3          186   33.92867
#> 74     3          186   37.50149
#> 75     3          186   34.41154
#> 76     3          186   30.85715
#> 77     3          186   37.74418
#> 78     4          198   37.27254
#> 79     4          198   35.19389
#> 80     4          198   33.74776
#> 81     4          198   29.70056
#> 82     4          198   38.98713
#> 83     4          198   29.82967
#> 84     4          198   42.84773
#> 85     4          198   39.79017
#> 86     4          198   31.53216
#> 87     4          198   27.83950
#> 88     4          198   29.31528
#> 89     4          198   33.83253
#> 90     4          198   31.48083
#> 91     4          198   31.00986
#> 92     4          198   28.18882
#> 93     4          198   32.42260
#> 94     4          198   28.96738
#> 95     4          198   24.84359
#> 96     4          198   30.85722
#> 97     4          198   36.92460
#> 98     4          198   29.94601
#> 99     4          198   35.47207
#> 100    4          198   25.07737
#> 101    4          198   32.37341
#> 102    4          198   35.05851
#> 103    5          210   34.45186
#> 104    5          210   33.53898
#> 105    5          210   30.05337
#> 106    5          210   29.07735
#> 107    5          210   28.26279
#> 108    5          210   33.59488
#> 109    5          210   28.62077
#> 110    5          210   30.75457
#> 111    5          210   31.84952
#> 112    5          210   41.65631
#> 113    5          210   30.00087
#> 114    5          210   34.14473
#> 115    5          210   33.40955
#> 116    5          210   28.55360
#> 117    5          210   32.71246
#> 118    5          210   39.79152
#> 119    5          210   35.15400
#> 120    5          210   33.23803
#> 121    5          210   31.07241
#> 122    5          210   23.45681
#> 123    5          210   38.32882
#> 124    5          210   26.22428
#> 125    5          210   36.50103
#> 126    5          210   41.96099
#> 127    5          210   26.30249
#> 128    6          222   36.97645
#> 129    6          222   32.47465
#> 130    6          222   26.35720
#> 131    6          222   26.62562
#> 132    6          222   26.21994
#> 133    6          222   31.21978
#> 134    6          222   26.87272
#> 135    6          222   36.91169
#> 136    6          222   43.50663
#> 137    6          222   27.68868
#> 138    6          222   37.37786
#> 139    6          222   37.29054
#> 140    6          222   35.25050
#> 141    6          222   28.99000
#> 142    6          222   33.14127
#> 143    6          222   32.38967
#> 144    6          222   36.32828
#> 145    6          222   31.95983
#> 146    6          222   38.26158
#> 147    6          222   31.94982
#> 148    6          222   38.61528
#> 149    6          222   28.79946
#> 150    6          222   27.81419
#> 151    6          222   48.83477
#> 152    6          222   31.75239
#> 153    6          222   35.09184
#> 154    6          222   36.67190
#> 155    6          222   31.43986
#> 156    6          222   36.11286
#> 157    6          222   35.42218
#> 158    6          222   32.69329
#> 159    6          222   34.00403
#> 160    6          222   33.54002
#> 161    7          234   44.63233
#> 162    7          234   31.23042
#> 163    7          234   29.57416
#> 164    7          234   34.86894
#> 165    7          234   36.14241
#> 166    7          234   36.73103
#> 167    7          234   32.55190
#> 168    7          234   29.72673
#> 169    7          234   40.59154
#> 170    7          234   33.05960
#> 171    7          234   30.65052
#> 172    7          234   33.58904
#> 173    7          234   33.77165
#> 174    7          234   39.87579
#> 175    7          234   35.08819
#> 176    7          234   38.21390
#> 177    7          234   32.36077
#> 178    7          234   35.69392
#> 179    7          234   33.17618
#> 180    7          234   35.13417
#> 181    7          234   30.51112
#> 182    7          234   28.57102
#> 183    7          234   44.01945
#> 184    7          234   37.49777
#> 185    7          234   28.84903
#> 186    7          234   31.83832
#> 187    7          234   29.15627
#> 188    8          246   46.37898
#> 189    8          246   42.23950
#> 190    8          246   34.87231
#> 191    8          246   38.64725
#> 192    8          246   34.17557
#> 193    8          246   33.88646
#> 194    8          246   32.42776
#> 195    8          246   33.33367
#> 196    8          246   43.82027
#> 197    8          246   35.85822
#> 198    8          246   36.66741
#> 199    8          246   37.24856
#> 200    8          246   41.86620
#> 201    8          246   33.70052
#> 202    8          246   31.47553
#> 203    8          246   43.93604
#> 204    8          246   34.05030
#> 205    8          246   32.73382
#> 206    9          258   32.19075
#> 207    9          258   31.96452
#> 208    9          258   35.28369
#> 209    9          258   40.85014
#> 210    9          258   43.14713
#> 211    9          258   41.26858
#> 212    9          258   36.26586
#> 213    9          258   38.24317
#> 214    9          258   34.67368
#> 215    9          258   34.61473
#> 216    9          258   42.09546
#> 217    9          258   33.22132
#> 218    9          258   47.09536
#> 219    9          258   37.54235
#> 220    9          258   38.96604
#> 221    9          258   34.51522
#> 222    9          258   35.28175
#> 223    9          258   31.81368
#> 224    9          258   37.10988
#> 225    9          258   39.92079
#> 226    9          258   39.47864
#> 227   10          270   36.47684
#> 228   10          270   36.44375
#> 229   10          270   37.78135
#> 230   10          270   47.11322
#> 231   10          270   34.81541
#> 232   10          270   39.29044
#> 233   10          270   49.01064
#> 234   10          270   39.65506
#> 235   10          270   33.77616
#> 236   10          270   37.02214
#> 237   10          270   42.39371
#> 238   10          270   38.37255
#> 239   10          270   37.50265
#> 240   10          270   38.52052
#> 241   10          270   40.54923
#> 242   10          270   47.59165
#> 243   10          270   39.71301
#> 244   10          270   45.17395
#> 245   10          270   43.07224
#> 246   10          270   39.59592
#> 247   10          270   32.96796
#> 248   10          270   37.69300
#> 249   11          282   40.05093
#> 250   11          282   42.55883
#> 251   11          282   48.41055
#> 252   11          282   53.04730
#> 253   11          282   49.56583
#> 254   11          282   41.71681
#> 255   11          282   34.13564
#> 256   11          282   40.52302
#> 257   11          282   42.75522
#> 258   11          282   46.28483
#> 259   11          282   46.83363
#> 260   11          282   45.53435
#> 261   11          282   35.82269
#> 262   11          282   46.30646
#> 263   11          282   40.25320
#> 264   11          282   43.15495
#> 265   11          282   42.68678
#> 266   11          282   44.33816
#> 267   11          282   42.45386
#> 268   11          282   34.55151
#> 269   11          282   45.77806
#> 270   11          282   44.14137
#> 271   11          282   41.09803
#> 272   11          282   42.89036
#> 273   11          282   42.96458
#> 274   11          282   43.37078
#> 275   11          282   50.00137
#> 276   11          282   41.31566
#> 277   11          282   43.12349
#> 278   11          282   47.79498
#> 279   11          282   47.26165
#> 280   12          294   49.66535
#> 281   12          294   41.62020
#> 282   12          294   53.66857
#> 283   12          294   44.62846
#> 284   12          294   53.03517
#> 285   12          294   38.00826
#> 286   12          294   44.41496
#> 287   12          294   50.15407
#> 288   12          294   40.97679
#> 289   12          294   40.80191
#> 290   12          294   39.93400
#> 291   12          294   39.40173
#> 292   12          294   42.27544
#> 293   12          294   45.86358
#> 294   12          294   34.91061
#> 295   12          294   45.30692
#> 296   12          294   50.09224
#> 297   12          294   53.83244
#> 298   12          294   50.39346
#> 299   12          294   47.85111
#> 300   12          294   36.25314
#> 301   12          294   41.50793
#> 302   12          294   42.67291
#> 303   12          294   47.60243
#> 304   12          294   43.82349
#> 305   12          294   38.43908
#> 306   13          306   53.75113
#> 307   13          306   50.14101
#> 308   13          306   46.99361
#> 309   13          306   51.57338
#> 310   13          306   39.63273
#> 311   13          306   48.97084
#> 312   13          306   43.44281
#> 313   13          306   49.07790
#> 314   13          306   45.60077
#> 315   13          306   48.84074
#> 316   13          306   52.12168
#> 317   13          306   45.91886
#> 318   13          306   50.63681
#> 319   13          306   40.33482
#> 320   13          306   42.51492
#> 321   13          306   52.97956
#> 322   13          306   47.64721
#> 323   13          306   36.30093
#> 324   13          306   39.51476
#> 325   13          306   44.94716
#> 326   13          306   49.92800
#> 327   13          306   45.40902
#> 328   13          306   48.79977
#> 329   13          306   50.36337
#> 330   13          306   53.68864
#> 331   13          306   46.14641
#> 332   13          306   45.64206